+++ research-app-messages/src/layers.ts
@@ -71,6 +71,12 @@ export interface LoadHipsCatalogCompletedMessage {
   spreadsheetInfo: SpreadSheetLayerInfo;
 }
 
+export function isLoadHipsCatalogCompletedMessage(o: any): o is LoadHipsCatalogCompletedMessage {  // eslint-disable-line @typescript-eslint/no-explicit-any
+  return o.event == 'layer_hipscat_load_completed' &&
+    typeof o.threadId == 'string' &&
+    typeof o.spreadsheetInfo == 'object';
+}
+
 /** Information about a "spreadsheet" layer in the engine.
  *
  * This interface is used when a client needs to learn about a spreadsheet layer
@@ -135,4 +141,79 @@ export interface GetHipsCatalogDataInViewReply {
    * hitting a limit on the amount of data to be returned at once.
    * */
   aborted: boolean;
+}
+
+/** A command to modify multiple table layer settings in a generic way.
+ * 
+ * This interface does not validate the setting names or their values.
+ */
+export interface MultiModifyTableLayerMessage {
+  /** The tag identifying this message type */
+  event: "table_layer_set_multi";
+
+  /** The identifier of the layer to modify */
+  id: string;
+
+  /** The names of the settings to modify */
+  settings: string[];
+
+  /** The values for these settings, in the same order as the setting names */
+  values: any[];  // eslint-disable-line @typescript-eslint/no-explicit-any
+}
+
+export function isMultiModifyTableLayerMessage(o: any): o is MultiModifyTableLayerMessage {  // eslint-disable-line @typescript-eslint/no-explicit-any
+  return o.event === 'table_layer_set_multi' &&
+    typeof o.id === 'string' &&
+    typeof o.settings === 'object' &&
+    typeof o.values === 'object';
+}
+
+/** A command to modify multiple FITS layer settings in a generic way.
+ * 
+ * This interface does not validate the setting names or their values.
+ */
+export interface MultiModifyFitsLayerMessage {
+  /** The tag identifying this message type */
+  event: "image_layer_set_multi";
+
+  /** The identifier of the IFTS layer to modify */
+  id: string;
+
+  /** The names of the settings to modify */
+  settings: string[];
+
+  /** The values for these settings, in the same order as the setting names */
+  values: any[];  // eslint-disable-line @typescript-eslint/no-explicit-any
+}
+
+export function isMultiModifyFitsLayerMessage(o: any): o is MultiModifyFitsLayerMessage {  // eslint-disable-line @typescript-eslint/no-explicit-any
+  return o.event === 'image_layer_set_multi' &&
+    typeof o.id === 'string' &&
+    typeof o.settings === 'object' &&
+    typeof o.values === 'object';
+}
+
+/** A command to modify multiple annotation settings in a generic way
+ * 
+ * This interface does not validate the setting names or their values.
+ */
+export interface MultiModifyAnnotationMessage {
+  /** The tag identifying this message type */
+  event: "annotation_set_multi";
+
+  /** The identifier of the annotation to modify */
+  id: string;
+
+  /** The names of the settings to modify */
+  settings: string[];
+
+  /** The values for these settings, in the same order as the setting names */
+  values: any[];  // eslint-disable-line @typescript-eslint/no-explicit-any
+}
+
+export function isMultiModifyAnnotationMessage(o: any): o is MultiModifyAnnotationMessage {  // eslint-disable-line @typescript-eslint/no-explicit-any
+  return o.event === 'annotation_set_multi' &&
+    typeof o.id === 'string' &&
+    typeof o.settings === 'object' &&
+    typeof o.values === 'object';
 }